chore: Rename and move exact_match function to metrics
